refactor(runme-api): specific calls to a single file
