+++ src/Helpers/index.js
@@ -3,6 +3,6 @@ export * from './EnvVars'
 export * from './QueryParam'
 export * from './SpecGenerator'
 export * from './Validation'
-export * from './WebSocket'
 export * from './Modal'
 export * from './copyToClipboard'
+export * from './RunmeApi'
